refactor: organize web tests by service
Similar to the refactor for node.js, we organize the web integration
tests by service.
