+++ packages/client-sdk-web/package.json
@@ -16,9 +16,13 @@
     "prebuild": "eslint . --ext .ts",
     "test": "jest --testPathIgnorePatterns auth-client.test.ts --maxWorkers 1",
     "unit-test": "jest unit",
-    "integration-test-auth": "jest --env=jsdom auth-client.test.ts --maxWorkers 1",
-    "integration-test-leaderboard": "jest --env=jsdom leaderboard --maxWorkers 1",
-    "integration-test-store": "jest --env=jsdom storage.test.ts --maxWorkers 1",
+    "integration-test-auth": "jest --env=jsdom auth/ --maxWorkers 1",
+    "integration-test-cache": "jest --env=jsdom cache/ --maxWorkers 1",
+    "integration-test-control-cache-topics": "npm run integration-test-cache && npm run integration-test-topics",
+    "integration-test-leaderboard": "jest --env=jsdom leaderboard/ --maxWorkers 1",
+    "integration-test-storage": "jest --env=jsdom storage/ --maxWorkers 1",
+    "integration-test-store": "npm run integration-test-storage",
+    "integration-test-topics": "jest --env=jsdom topics/ --maxWorkers 1",
     "integration-test-jsdom": "jest integration --env=jsdom --testMatch \"**/dictionary.test.ts|**/ping.test.ts|*/topic-client.test.ts|leaderboard.test.ts\" --maxWorkers 1",
     "integration-test-happy-dom": "jest integration --env=@happy-dom/jest-environment --testPathIgnorePatterns \"dictionary.test.ts|ping.test.ts|topic-client.test.ts|auth-client.test.ts|leaderboard.test.ts|storage.test.ts\" --maxWorkers 1",
     "integration-test": "npm run integration-test-happy-dom && npm run integration-test-jsdom",
