+++ packages/client-sdk-web/test/integration/shared/auth/auth-client.test.ts
@@ -1,5 +1,5 @@
 import {runAuthClientTests} from '@gomomento/common-integration-tests';
-import {SetupAuthClientIntegrationTest} from '../integration-setup';
+import {SetupAuthClientIntegrationTest} from '../../integration-setup';
 
 const {
   mgaAccountSessionTokenAuthClient,
