+++ packages/client-sdk-web/test/integration/shared/cache/batch-get-set.test.ts
@@ -1,5 +1,5 @@
 import {runBatchGetSetTests} from '@gomomento/common-integration-tests';
-import {SetupIntegrationTest} from '../integration-setup';
+import {SetupIntegrationTest} from '../../integration-setup';
 
 const {cacheClient, cacheClientWithThrowOnErrors, integrationTestCacheName} =
   SetupIntegrationTest();
