+++ packages/client-sdk-web/test/integration/shared/cache/get-set-delete.test.ts
@@ -1,5 +1,5 @@
 import {runGetSetDeleteTests} from '@gomomento/common-integration-tests';
-import {SetupIntegrationTest} from '../integration-setup';
+import {SetupIntegrationTest} from '../../integration-setup';
 
 const {
   cacheClient,
