+++ packages/client-sdk-web/test/integration/shared/cache/list.test.ts
@@ -1,5 +1,5 @@
 import {runListTests} from '@gomomento/common-integration-tests';
-import {SetupIntegrationTest} from '../integration-setup';
+import {SetupIntegrationTest} from '../../integration-setup';
 
 const {cacheClient, integrationTestCacheName} = SetupIntegrationTest();
