+++ packages/client-sdk-web/test/integration/shared/storage/storage.test.ts
@@ -1,5 +1,5 @@
 import {runStorageServiceTests} from '@gomomento/common-integration-tests';
-import {SetupStorageIntegrationTest} from '../integration-setup';
+import {SetupStorageIntegrationTest} from '../../integration-setup';
 
 const {storageClient, integrationTestStoreName} = SetupStorageIntegrationTest();
