+++ packages/client-sdk-web/test/integration/shared/topics/topic-client.test.ts
@@ -1,5 +1,5 @@
 import {runTopicClientTests} from '@gomomento/common-integration-tests';
-import {SetupTopicIntegrationTest} from '../integration-setup';
+import {SetupTopicIntegrationTest} from '../../integration-setup';
 
 const {
   topicClient,
